home コンポーネントの追加
