+++ src/App.tsx
@@ -1,50 +1,15 @@
 import './App.css';
-import Header from './components/Header/Header';
-import Hero from './components/Hero/Hero';
-import Features from './components/Features/Features';
-import Testimonials from './components/Testimonials/Testimonials';
-import Contact from './components/Contact/Contact';
-import Footer from './components/Footer/Footer';
-import { useState } from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+
+import Home from './components/Home/Home';
 
 function App() {
-  const features = [
-    {
-      title: "最高品質",
-      description: "厳選された素材と職人技が生み出す、比類なき品質をお届けします。"
-    },
-    {
-      title: "使いやすさ",
-      description: "直感的なデザインで、誰もが簡単に使いこなせる製品です。"
-    },
-    {
-      title: "24時間サポート",
-      description: "お客様のニーズに合わせて、いつでもどこでもサポートいたします。"
-    },
-  ]
-
-  const [darkMode, setDarkMode] = useState(false);
-
-  const changeDarkMode = () => {
-    setDarkMode(!darkMode);
-  }
-
   return (
-    <div className={`App ${darkMode ? 'dark' : 'light'}`}>
-      <Header changeDarkMode={changeDarkMode} />
-
-      <main>
-        <Hero />
-
-        <Features title="特徴" features={features} />
-
-        <Testimonials />
-
-        <Contact />
-      </main>
-
-      <Footer />
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/" element={<Home />} />
+      </Routes>
+    </Router>
   );
 }
